Proposal to add `oci` as purl type
The current purl specification has a proposed `Docker` entry that can be
used when referencing docker container images. However, Docker Hub is
not the only registry where container images might be stored.
Additionally, a container image is not the only type of artifact that
may be stored in an oci registry. This commit adds an `oci` package type
and points users to use it instead of `docker` for container images.

In order to represent all cloud native artifacts, there needs to be an
entry in the purl spec that can represent more than just docker
container images. Adding the `oci` entry would specifically
enable:

- a means to uniquely identify any artifact stored in an OCI-compliant
registry, regardless of what registry it may currently be in. (Use
cases: importing an artifact from a public registry to a private
registry, and promotion within and across private registries)

Signed-off-by: Rose Judge <rjudge@vmware.com>
Signed-off-by: Nisha Kumar <nishak@vmware.com>
Signed-off-by: Steve Laskar <steve.lasker@microsoft.com>