add an example with `/` in namespace to README and SPEC
I think it is not clear from the text in the README and the examples
whether a `/` in the namespace needs to be escaped. This adds an example
from `/test-suite-data.json` to the list of examples, to make that
clear, that escaping should not be done.

A follow-up question, that is not scope of this PR, would be:
=============================================================
given the PURL `pkg:swift/github.com%2FAlamofire/Alamofire@5.4.3`
is `pkg:swift/github.com/Alamofire/Alamofire@5.4.3` its canonical form?
If yes, it should be added to the test cases.

Or, for having more fun and with looking at #63 : what is the canonical
form of `pkg:golang/github.com%2Frussross%2Fblackfriday%2Fv2@v2.1.0`?

Signed-off-by: Maximilian Huber <gh@maxhbr.de>