Removes case-sensitivity for golang names
Full Go Module names must be case-sensitive because git (one of the supported VCSs) is case-sensitive. 

For instance, if you change the casing a module in your go.mod file, go will complain on "go build":

        module declares its path as: github.com/foo/bar
                but was required as: github.com/FOO/BAR

GitHub treats them the same, but only because it was built with case-insensitive urls. Other git repositories may not be the same.