Update PURL-TYPES.rst with Mendix package type
The Mendix platform is a low-code platform with its own package system. Mendix packages are imported into Mendix Apps, or used as IDE extensions and hosted on the Mendix Marketplace , see https://marketplace.mendix.com/. A SBOM is created for the App which used package urls to list the components in the App.