Remove section on Character Encoding
This section contained duplicate information that is already present for
each component in the "Rules" section and included confusing sentences
that applied too generally to certain purl components.

Each component's set of rules should clearly define how the content of
that component should (or should not) be encoded. These improvements are
being made in PRs such as #383 (more will be coming for other component
sections).